requirements.txt
amqp==2.5.0
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.3
django-admin-rangefilter==0.5.0
django-bootstrap4==0.0.8
django-class==2019.4.13
django-crispy-forms==1.7.2
django-filter==2.2.0
django-heroku==0.3.1
django-import-export==1.2.0
django-jet==1.0.8
djangorestframework==3.9.4
et-xmlfile==1.0.1
google-api-python-client==1.4.1
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
jdcal==1.4.1
kombu==4.6.3
Markdown==3.1.1
meld3==1.0.2
oauth2client==4.1.3
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
psutil==5.6.3
psycopg2==2.8.3
public==2019.4.13
pyasn1==0.4.6
pyasn1-modules==0.2.6
pyte==0.8.0
pytz==2019.1
PyYAML==5.1.1
redis==3.3.6
requests==2.22.0
rsa==4.0
sentry-sdk==0.10.2
six==1.12.0
sqlparse==0.3.0
supervisor==4.0.4
tablib==0.13.0
thefuck==3.29
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
whitenoise==4.1.3
xlrd==1.2.0
xlwt==1.3.0
